+++ src/location/mod.rs
@@ -11,6 +11,8 @@ pub trait CurrentLocation {
 #[derive(Default, Debug, Serialize, Deserialize)]
 pub struct Location {
     #[serde(default)]
+    pub city: String,
+    #[serde(default)]
     pub lat: f32,
     #[serde(default)]
     pub lon: f32,